Game 1 - Slime Crime
Simple shooter with infinite wave enemies that progressively increase in speed and difficulty. The turret reloads faster as the waves progress to match the slimes growing speed and the addition of Boss Slimes that are harder to kill. I decided to not implement any further enemy functionality since I have previously designed slime based enemies and didn't wanna reuse the concepts for this one.